Backport .skyd file fixes from develop (#756)
* Fix missing chromatograms in .skyd files whose format is older than 12 (#753)

* Fix "Unable to find a matching chromatogram" error with Skyline documents with .skyd format earlier than 12 (Skyline documents from before 2017).

* Support for skyd file format 18 needed for Skyline 23.1 (#742)

Skyline 23.1 will have some changes to the .skyd file format which will need to be supported by Panorama

Changes:
The layout of the structure "ChromGroupHeaderInfo" changed a lot in the latest .skyd file format, so there's a big "if" branch in the ChromGroupHeaderInfo constructor that reads the fields in different order.
Instead of just having a string for each chromatogram TextId, there is a protocol buffer called "ChromatogramGroupId" which has separate fields for the peptide modified sequence and all the small molecule fields.
The reason that ChromatogramGroupId was introduced in Skyline is so that precursors with different "spectrum filters" can be distinguished in the .skyd file. This pull request currently just ignores any spectrum filters which might be in the .skyd file.